+++ documentation/manual/scalaGuide/main/sql/ScalaAnorm.md
@@ -239,7 +239,26 @@ SQL("Select name,image from Country")().map {
 
 ### Database interoperability
 
-Note that different databases will return different data types in the Row. For instance, an SQL 'smallint' is returned as a Short by org.h2.Driver and an Integer by org.postgresql.Driver. A solution to this is to simply write separate case statements for each database (i.e. one for development and one for production).
+Note that different databases will return different data types in the Row. For instance, an SQL 'smallint' is returned as a Short by `org.h2.Driver` and an Integer by `org.postgresql.Driver`. A solution to this is to simply write separate case statements for each database (i.e. one for development and one for production).
+
+Anorm provides common mappings for Scala types from JDBC datatypes.
+
+When needed, it's possible to customize such mappings, for example if underlying DB doesn't support boolean datatype and returns integer instead. To do so, you have to provide a new implicit conversion for `Column[T]`, where `T` is the target Scala type:
+
+```scala
+import anorm.Column
+
+// Custom conversion from JDBC column to Boolean
+implicit def columnToBoolean: Column[Boolean] = 
+  Column.nonNull { (value, meta) =>
+    val MetaDataItem(qualified, nullable, clazz) = meta
+    value match {
+      case bool: Boolean => Right(bool) // Provided-default case
+      case bit: Int      => Right(bit == 1) // Custom conversion
+      case _             => Left(TypeDoesNotMatch(s"Cannot convert $value: ${value.asInstanceOf[AnyRef].getClass} to Boolean for column $qualified"))
+    }
+  }
+```
 
 ## Dealing with Nullable columns
 
@@ -459,27 +478,6 @@ $ spokenLanguages("FRA")
         Arabic, Italian, Portuguese, Spanish, Turkish
     ))
 )
-```
-
-### Custom column mapping
-
-Anorm provides common mappings for Scala types from JDBC datatypes.
-
-When needed, it's possible to customize such mappings, for example if underlying DB doesn't support boolean datatype and returns integer instead. To do so, you have to provide a new implicit conversion for `Column[T]`, where `T` is the target Scala type:
-
-```scala
-import anorm.Column
-
-// Custom conversion from JDBC column to Boolean
-implicit def columnToBoolean: Column[Boolean] = 
-  Column.nonNull { (value, meta) =>
-    val MetaDataItem(qualified, nullable, clazz) = meta
-    value match {
-      case bool: Boolean => Right(bool) // Provided-default case
-      case bit: Int      => Right(bit == 1) // Custom conversion
-      case _             => Left(TypeDoesNotMatch(s"Cannot convert $value: ${value.asInstanceOf[AnyRef].getClass} to Boolean for column $qualified"))
-    }
-  }
 ```
 
 > **Next:** [[Integrating with other database access libraries | ScalaDatabaseOthers]]
