+++ framework/src/anorm/src/main/scala/anorm/package.scala
@@ -1,4 +1,3 @@
-
 /*
  * Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
  */
